% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanFilename.R
\name{cleanFilename}
\alias{cleanFilename}
\title{cleanFilename}
\usage{
cleanFilename(f)
}
\arguments{
\item{f}{Complete file path}
}
\value{
Clean file name without "/" and the file extension.
}
\description{
Removes the extension and path in a file name f.
}
\seealso{
trimextcsv
}
\keyword{trimming}
