% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeSW.R
\name{completeSW}
\alias{completeSW}
\title{completeSW}
\usage{
completeSW(sw, win)
}
\arguments{
\item{sw}{dataframe with all superwindows}

\item{win}{dataframe with all windows}
}
\description{
Calls the function reduceWin, passing as arguments the superwindows 
and windows within 1 sequence each time, so that superwindow data 
are filled. Returns a data frame with all superwindows with complete data
(start, end, chrom, mask, pasted string with windowIDs).
}
\seealso{
reduceWin
}
\keyword{superwindows}
\keyword{windows,}
