% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMatches.R
\name{filterMatches}
\alias{filterMatches}
\title{filterMatches}
\usage{
filterMatches(views, pattern = "")
}
\arguments{
\item{views}{an object of the class XStringViews}

\item{pattern}{a string of characters that contains the motif core
pattern (e.g.: "[AT]CAA") to filter the matrix alignments. Accepts "" 
as a pattern.}
}
\value{
A dataframe with the matches (all the 'views' objects in it).
}
\description{
Filters sequences that do not match a pattern.
}
\keyword{filters,}
\keyword{matches}
\keyword{sequences,}
