% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterseqsbyDist.R
\name{filterseqsbyDist}
\alias{filterseqsbyDist}
\title{filterseqsbyDist}
\usage{
filterseqsbyDist(seqs, dist, min_width = 0)
}
\arguments{
\item{seqs}{input sequences dataframe}

\item{dist}{integer vector of length 2; the first one is the distance
to the first exon allowed for the promoter (<= 0) and the second one 
is the maximum distance of introns (>= 0)}

\item{min_width}{Minimum sequence length allowed (int). Defaults to 0.}
}
\value{
A dataframe object equivalent to the input seqs.
}
\description{
Takes a dataframe with all the sequences and reduces them to a maximum 
distance from the first exon start. It also allows to filter out sequences
with length <= min_width.
}
\seealso{
trimSequencetoDist
}
\keyword{dist}
\keyword{filters,}
\keyword{seqs,}
