% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSWbyDist.R
\name{filterSWbyDist}
\alias{filterSWbyDist}
\title{filterSWbyDist}
\usage{
filterSWbyDist(sw, dist, include = TRUE)
}
\arguments{
\item{sw}{dataframe with superwindows (intervals). 
Must have a "dist_to_first_exon" column.}

\item{dist}{numeric vector of length 2. The first element is the maximum 
upstream distance; the second element is the maximum downstream distance.}

\item{include}{if TRUE, includes partially overlapping sw and trims "width" 
to the overlapping region. If FALSE, omits overlapping windows (defaults
to TRUE)}
}
\value{
A dataframe with the same columns as the superwindows dataframe.
}
\description{
Filters out the superwindows which distances to the first gene exon
are not within a given interval (parameter dist).
}
\keyword{distance,}
\keyword{filters,}
\keyword{superwindows}
