% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiltFromPWMs.R
\name{getFiltFromPWMs}
\alias{getFiltFromPWMs}
\title{getFiltFromPWMs}
\usage{
getFiltFromPWMs(PWMs)
}
\arguments{
\item{PWMs}{list of matrixes, with names}
}
\value{
A vector of strings, with names.
}
\description{
Given a set of PWMs, produces a set of strings representing each PWM 
(a regular expression allowing at each position all the nucleotides 
with probability > 0).
}
\keyword{PWMs,}
\keyword{filters,}
\keyword{strings}
