% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGCProp.R
\name{getGCProp}
\alias{getGCProp}
\title{getGCProp}
\usage{
getGCProp(seq)
}
\arguments{
\item{seq}{an input character sequence}
}
\value{
The proportion of GCs (range 0-1).
}
\description{
Calculates the GC content of a character sequence.
}
\keyword{GC,}
\keyword{strings}
