% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneStats.R
\name{getGeneStats}
\alias{getGeneStats}
\title{getGeneStats}
\usage{
getGeneStats(seqs, allsuperwindows, typicalName = TRUE)
}
\arguments{
\item{seqs}{dataframe with original sequences data}

\item{allsuperwindows}{dataframe with superwindow data. Must have
"seqID", "width", "width_withoutreps"... columns (gene column no needed)}

\item{typicalName}{whether the seqIDs are in the form 'genename_seqnumber'
or not. If FALSE, the "seqID" column is used to split the data instead 
of the "gene" column (defaults to TRUE).}
}
\value{
A dataframe with basic stats calculated for each gene.
}
\description{
This function gets basic stats for each gene.
}
\keyword{gene,}
\keyword{stats}
