% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifs.R
\name{getMotifs}
\alias{getMotifs}
\title{getMotifs}
\usage{
getMotifs(seqstr, mymotifs, myfilt, minscore)
}
\arguments{
\item{seqstr}{the sequence where the motifs are to be aligned}

\item{mymotifs}{a list of PWMatrixes (the ones that are to be aligned 
to seq), each with a name}

\item{myfilt}{a character vector with the motif core pattern 
(e.g "[AT]CAA") to filter the matrix alignments}

\item{minscore}{is the min.score argument for Biostrings::matchPWM function}
}
\value{
A dataframe with all the matches present in both the + and - strand
}
\description{
Searches motifs in a string of characters, filters the ones that 
don't have the core (myfilt).
}
\seealso{
filterMatches, reverseOffsets
}
\keyword{filters,}
\keyword{motifs,}
\keyword{strings}
