% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPartitions.R
\name{getPartitions}
\alias{getPartitions}
\title{getPartitions}
\usage{
getPartitions(n, k)
}
\arguments{
\item{n}{number of tasks}

\item{k}{number of groups}
}
\description{
Divides n tasks into k groups. Used to split datasets and run in parallel.
It returns a list in which each element is a vector with the indexes (from 1
to n) in each partition
}
\keyword{parallel}
\keyword{partitions,}
