% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomPWMs.R
\name{getRandomPWMs}
\alias{getRandomPWMs}
\title{getRandomPWMs}
\usage{
getRandomPWMs(PWMs, n = 10)
}
\arguments{
\item{PWMs}{list of matrixes, with names}

\item{n}{number of permutated lists (defaults to n=10).}
}
\value{
A list of lists, with n permutations of each PWM (1 shuffled PWM
per originl PWM)
}
\description{
Given a set of PWMs, returns n lists of PWMs with shuffled columns.
}
\keyword{PWMs}
