% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomPWMsAndFilts.R
\name{getRandomPWMsAndFilts}
\alias{getRandomPWMsAndFilts}
\title{getRandomPWMsAndFilts}
\usage{
getRandomPWMsAndFilts(PWMs, filt, n = 10)
}
\arguments{
\item{PWMs}{list of matrixes, with names}

\item{filt}{list of filters (regular expression-like, eg. [AT]TTG[CG]A)}

\item{n}{number of permutated lists (defaults to n=10).}
}
\value{
n lists of PWMs with shuffled columns (1 shuffled PWM per 
originalPWM) and its corresponding filters (a list of lists, with n 
permutations of each PWM.)
}
\description{
Given a set of PWMs and filter (eg. [AT]TTG[CG]A), returns n lists of PWMs
with shuffled columns and its corresponding filters.
}
\keyword{PWMs,}
\keyword{filters}
