% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSWdistFromATG.R
\name{getSWdistFromATG}
\alias{getSWdistFromATG}
\title{getSWdistFromATG}
\usage{
getSWdistFromATG(start, end, strand, region, seqdist, seqwidth)
}
\arguments{
\item{start}{interval start within original sequence 
(not chromosomal coordinate)}

\item{end}{interval end within original sequence}

\item{strand}{interval strand (equal to sequence strand)}

\item{region}{"upstream" or "inton"}

\item{seqdist}{distance of original sequence to first exon start}

\item{seqwidth}{original sequence width}
}
\value{
An integer number with the distance (in bp).
}
\description{
Calculates superwindows distance to first exon (not necessarily ATG) 
counting from the end of the superwindow (not the beginning!).
}
\keyword{distance}
\keyword{superwindows,}
