% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSWSeq.R
\name{getSWSeq}
\alias{getSWSeq}
\title{getSWSeq}
\usage{
getSWSeq(start, end, mask, originalseq, maskedseq)
}
\arguments{
\item{start}{interval start}

\item{end}{interval end}

\item{mask}{"original" if not masked, otherwise the masked sequence 
will be used}

\item{originalSeq}{raw DNA sequence (character string)}

\item{maskedSeq}{masked DNA sequence (usually repeats have been 
replaced by '+') (character string)}
}
\value{
A character string with the subsequence found.
}
\description{
Given a sequence and 2 coordinates, returns the subsequence within 
the two coordinates, masked or unmasked, depending on the parameter "mask".
}
\keyword{strings}
\keyword{superwindows,}
