% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWindowCond.R
\name{getWindowCond}
\alias{getWindowCond}
\title{getWindowCond}
\usage{
getWindowCond(window_motifs, nkinds = 6, PWMsubset)
}
\arguments{
\item{window_motifs}{data frame with motifs. Column "motif_name" 
is compared to PWMsubset}

\item{nkinds}{minimum number of different motifs that a window needs 
to have (in case PWMsubset is null). Defaults to 6.}

\item{PWMsubset}{character vector with motifs that are necessary to 
form a window}
}
\value{
If aa set of motifs is given (PWMsubset), returns TRUE if all 
the motifs names in PWMsubset are found in window_motifs. If, otherwise, 
PWMsubset is NULL, it returns TRUE if there are at least nkinds 
kinds of motifs.
}
\description{
Given a set of motifs (within a genomic range), it calculates whether 
they are enough to form a window or not.
}
\keyword{motifs}
\keyword{windows,}
