% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWindows.R
\name{getWindows}
\alias{getWindows}
\title{getWindows}
\usage{
getWindows(matches, maxwidth = 800, nkinds = 8, PWMsubset = NULL)
}
\arguments{
\item{matches}{data.frame with start, end, width, motif_names, 
strand, offsets}

\item{maxwidth}{maximum width of the window}

\item{nkinds}{number of different kinds of motif_names that need 
to be in a window}
}
\value{
A dataframe with the window data (start, end, and 
pasted motif information)
}
\description{
Given a dataframe with within-sequence coordinates for elements 
of different kinds, it gets all the windows of a maximum length 
maxwidth with at least nkinds different kinds of elements.
}
\keyword{motifs}
\keyword{windows,}
