% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalMakeWindows.R
\name{globalMakeWindows}
\alias{globalMakeWindows}
\title{globalMakeWindows}
\usage{
globalMakeWindows(seqs, motifList, seqfile, PWMset, PWMsubsetname,
  max_width, nmotifs = 6, PWMsubset = NULL)
}
\arguments{
\item{seqs}{dataframe with sequences. Must have "original.seq" and 
"masked.seq" columns as strings and "seqID" with unique identifiers.}

\item{motifList}{a list with data frames containing motif matches, as 
returned by globalSearchMotifs().}

\item{seqfile}{name of the set of sequences (the file from which seqs 
have been read, without extension). Will be used to write output files.}

\item{PWMset}{list of PWMs}

\item{PWMsubsetname}{a string with the name of PWMsubset (used to
name files). Usually is the first letter of each motif name, tied}

\item{max_width}{maximum window span (usually 700 bp)}

\item{nmotifs}{number of different motifs needed to make a window}

\item{PWMsubset}{vector with names of PWMs to use in search.}
}
\value{
A dataframe with all the information regarding windows. It also
writes both a .csv and a .bed file with the genomic intervals.
}
\description{
Produces a data frame with a row for every region of max_with bearing 
nmotifs different motifs. As input it uses motifList, which is the output 
of globalSearchMotifs() function.
}
\seealso{
globalSearchMotifs
}
\keyword{PWMs}
\keyword{motifs,}
\keyword{search,}
\keyword{windows,}
