% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalMakeWindowsPar.R
\name{globalMakeWindowsPar}
\alias{globalMakeWindowsPar}
\title{globalMakeWindowsPar}
\usage{
globalMakeWindowsPar(seqs, motifList, seqfile, PWMset, PWMsubsetname,
  max_width, nmotifs = 6, PWMsubset = NULL, n_partitions = 24)
}
\arguments{
\item{seqs}{data frame with sequences. Must have "original.seq" and 
"masked.seq" fields with strings and "seqID" with unique identifiers.}

\item{motifList}{a list with data frames containing motif matches, as 
returned by globalSearchMotifs().}

\item{seqfile}{name of the set of sequences (the file from which seqs 
have been read without extension). Will be used to write output files.}

\item{PWMset}{list of PWMs}

\item{PWMsubsetname}{a string with the name of the PWMsubset 
(used to name files), usually tie first letter of each motif name}

\item{max_width}{maximum window span (usually 700 bp)}

\item{nmotifs}{number of different motifs needed to make a window (6 
by default)}

\item{PWMsubset}{vector with names of PWMs to use in search.}

\item{n_partitions}{number of batches to split the data (24 by default)}
}
\value{
A dataframe with all the information regarding windows. It also
writes a .csv and a .bed file with the genomic intervals.
}
\description{
Splits the data in batches and calls globalMakeWindows in parallel.
}
\seealso{
getPartitions, globalMakeWindows, globalSearchMotifs
}
\keyword{PWMs,}
\keyword{motifs,}
\keyword{parallel}
\keyword{search,}
\keyword{windows,}
