% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalSearchMotifs.R
\name{globalSearchMotifs}
\alias{globalSearchMotifs}
\title{globalSearchMotifs}
\usage{
globalSearchMotifs(seqs, PWMsubset, mymotifs, myfilt, seqfile)
}
\arguments{
\item{seqs}{dataframe with sequences. Must have "original.seq" and
"masked.seq" columns as strings and "seqID" with unique identifiers}

\item{PWMsubset}{vector with names of PWMs to use in search.}

\item{mymotifs}{list of PWMs (matrixes). Must contain the PWMs in PWMsubset.}

\item{myfilt}{a string vector with regular expressions 
used to filter out PWM matches.}

\item{seqfile}{name of the set of sequences (the file from which seqs have
been read, without the extension). Will be used to write output files.}
}
\value{
A list with an element for every row of the originl sequences 
dataframe, each one containing all the motif matches in both the original 
and masked sequence. It also saves the list in a .RData file.
}
\description{
Searches a set of PWMs (including filtering) in a set of sequences 
and returns a list of lists, each one of them with a dataframe containing 
the matches in the original seq or the masked seq (masked repeats or 
other features). 
Be careful! before starting the search it will check whether a file exists 
with the same output name. If TRUE, it will load and won't perform
any search.
}
\seealso{
getMotifs
}
\keyword{PWMs}
\keyword{motifs,}
\keyword{search,}
