% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grangesFromDF.R
\name{grangesFromDF}
\alias{grangesFromDF}
\title{grangesFromDF}
\usage{
grangesFromDF(dataframe, seqnames = "seqID")
}
\arguments{
\item{dataframe}{input dataframe. Must contain variables named:
seqID, chrom_start, chrom_end, strand.}

\item{seqnames}{names for sequences (defaults to "seqID")}
}
\value{
A GenomicRanges object with the same intervals as the dataframe
}
\description{
Creates a GenomicRanges object from a data frame containing seqID, 
chrom_start, chrom_end and strand variables.
}
\keyword{granges}
