% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinSWFiles.R
\name{joinSWFiles}
\alias{joinSWFiles}
\title{joinSWFiles}
\usage{
joinSWFiles(swfiles, types = list(maxlen = "[2785]000*_", tfgroup =
  "EGhIS[pl]+"), rmark = "split[0-9]+", mark = "splitjoin")
}
\arguments{
\item{swfiles}{the superwindows files to be rejoined}

\item{types}{different regular expression to get the conditions}

\item{rmark}{how different splits are marked}

\item{mark}{used to label the output file with all splits together (will 
replace rmark in the name)}
}
\value{
A vector with the new swfiles.
}
\description{
When sequences have been split into many files, this function rejoins 
the superwindows files sorting them by condition.
}
\keyword{join,}
\keyword{superwindows}
