% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBoxplots.R
\name{makeBoxplots}
\alias{makeBoxplots}
\title{makeBoxplots}
\usage{
makeBoxplots(genestats, outnameaux)
}
\arguments{
\item{genestats}{stats generated from function getGeneStats}

\item{outnameaux}{name for the output PDF file}
}
\value{
Does not return anything; writes file with finalstats directly.
}
\description{
Makes boxplots from GeneStats. Creates a new directory ./boxplots
and saves the boxplots generated in PDF format.
}
\seealso{
getGeneStats
}
\keyword{boxplot}
\keyword{stats,}
