% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFinalStats.R
\name{makeFinalStats}
\alias{makeFinalStats}
\title{makeFinalStats}
\usage{
makeFinalStats(swfiles, seqs, dataset, makeplots = FALSE,
  dist_intervals = NULL, doDist = TRUE)
}
\arguments{
\item{swfiles}{files with superwindows (do not need to include group)}

\item{seqs}{dataframe with the original sequences analysed}

\item{dataset}{name of condition of the swfiles}

\item{makeplots}{if TRUE, makes boxplots of all variables (defaults to
FALSE)}

\item{dist_intervals}{calculates stats for intervals centered around TSS}

\item{doDist}{if TRUE, only considers 1 dist interval (-Inf, Inf). Defaults
to TRUE.}
}
\value{
A dataframe with the final stats.
}
\description{
Makes stats from superwindows. Reads superwindows file and:
1) calculates stats per gene
2) calculates mean and sum per gene
}
\seealso{
filterSWbyDist, filterseqsbyDist, getGeneStats, makeBoxplots
}
\keyword{boxplots,}
\keyword{graphs}
\keyword{stats,}
