% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFinalStats2.R
\name{makeFinalStats2}
\alias{makeFinalStats2}
\title{makeFinalStats2}
\usage{
makeFinalStats2(swfiles, seqs, dataset, makeplots = FALSE,
  typicalName = FALSE)
}
\arguments{
\item{swfiles}{files with superwindows (do not need to include group)}

\item{seqs}{dataframe with the original sequences analysed}

\item{dataset}{name of condition of the swfiles}

\item{makeplots}{if TRUE, makes boxplots of all variables (defaults to
FALSE)}

\item{typicalName}{whether the gene names are in the format 'gene_numseq' 
or not. If FALSE, "seqID" becomes "geneID" (defaults to FALSE).}
}
\value{
A dataframe with the final stats.
}
\description{
Makes stats from superwindows. Reads superwindows file and:
1) calculates stats per gene
2) calculates mean and sum per gene
}
\details{
Works better for genes which sequences are not divided in introns and
exons (it does not take into account dist_from_first_exon). Also, it
allows for absence of gene names.
}
\seealso{
getGeneStats, makeBoxplots
}
\keyword{boxplots,}
\keyword{graphs}
\keyword{stats,}
