% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSuperwindows.R
\name{makeSuperwindows}
\alias{makeSuperwindows}
\title{makeSuperwindows}
\usage{
makeSuperwindows(filename = "", seqs, outfolder = "./superwindows/",
  win = NULL)
}
\arguments{
\item{filename}{name of the windows file}

\item{seqs}{dataframe with the original sequences analysed}
}
\value{
Nothing, writes superwindows files directly.
}
\description{
Reads a windows file and calls all the functions that are required 
to make superwindows. As an output, writes superwindows files, both with 
and without sequences.
}
\seealso{
cleanFilename, getSuperwindows, completeSW, reduceWin, writeBedFile
}
\keyword{bed}
\keyword{superwindows,}
\keyword{windows,}
