% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSuperwindowsPar.R
\name{makeSuperwindowsPar}
\alias{makeSuperwindowsPar}
\title{makeSuperwindowsPar}
\usage{
makeSuperwindowsPar(filename = "", seqs, outfolder = "./superwindows/",
  win = NULL, n_partitions)
}
\arguments{
\item{filename}{name of the windows file}

\item{seqs}{dataframe with the original sequences analysed}

\item{n_partitions}{number of threads to use in parallel}
}
\value{
Nothing, writes superwindows files directly.
}
\description{
Reads a windows file and calls all the functions that are required 
to make superwindows (in parallel). As an output, writes superwindows 
files, both with and without sequences.
}
\seealso{
cleanFilename, getSuperwindows, completeSW, reduceWin, writeBedFile
}
\keyword{bed}
\keyword{superwindows,}
\keyword{windows,}
