% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMotifs.R
\name{mergeMotifs}
\alias{mergeMotifs}
\title{mergeMotifs}
\usage{
mergeMotifs(win)
}
\arguments{
\item{win}{window dataframe with the following fields:
"width", "signature", "offsets", "strands", "seqs", "starts", "ends".}
}
\value{
A dataframe with 1 row, containing the mentioned variables. 
Each field contains the all the motifs pasted into a single string, 
using "_" as separators.
}
\description{
Merges all the motif data from several windows into one row, avoiding
redundancies.
}
\keyword{merge,}
\keyword{motifs}
