% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeqFile.R
\name{readSeqFile}
\alias{readSeqFile}
\title{readSeqFile}
\usage{
readSeqFile(seqfile)
}
\arguments{
\item{seqfile}{the input sequences file}
}
\value{
The same file without possible missing data.
}
\description{
Reads a file with sequences and fills missing data to prevent 
further crashing.
}
\keyword{sequences,}
\keyword{strings}
