% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceWin.R
\name{reduceWin}
\alias{reduceWin}
\title{reduceWin}
\usage{
reduceWin(sw, win)
}
\arguments{
\item{sw}{the superwindow dataframe with superwindows sharing the same seqID}

\item{win}{the windows dataframe with windows sharing the same seqID}
}
\value{
A dataframe with all the superwindow data.
}
\description{
Takes data of several overlapping windows and calculates data for 
the superwindow: start, end, chromosome, seq.mask and a string with 
pasted windowIDs.
}
\keyword{superwindows}
\keyword{windows,}
