% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseOffsets.R
\name{reverseOffsets}
\alias{reverseOffsets}
\title{reverseOffsets}
\usage{
reverseOffsets(data, len, strand = "-")
}
\arguments{
\item{data}{dataframe with start, end, offsets, width and strand}

\item{len}{length of the original sequence where the subsequences in 
argument data were found}

\item{strand}{used to control it reverses only the negative strands}
}
\value{
A data frame with the same dimensions as the original. Start 
keeps being before end.
}
\description{
Calculates the offsets, starts and ends in the reverse complement
strand for the seqs of a given strand.
}
\keyword{complement}
\keyword{reverse,}
\keyword{strand,}
