% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimSequencetoDist.R
\name{trimSequencetoDist}
\alias{trimSequencetoDist}
\title{trimSequencetoDist}
\usage{
trimSequencetoDist(seq, dist)
}
\arguments{
\item{seq}{a row of a dataframe with the sequence data}

\item{dist}{integer vector of length 2. The first one number is the distance
to the first exon allowed for the promoter (<= 0) and the second one is 
the maximum distance of introns (>= 0).}
}
\value{
A dataframe row with the same format as seq, with trimmed data.
}
\description{
Gets the region of a sequence that is closer to the first exon than a 
given limit. Calculates start, end, and sequence (masked and original).
}
\keyword{distance}
\keyword{sequence,}
\keyword{trim,}
\keyword{trimming,}
