% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winToDataframe.R
\name{winToDataframe}
\alias{winToDataframe}
\title{winToDataframe}
\usage{
winToDataframe(window_motifs)
}
\arguments{
\item{window_motifs}{a dataframe with motif data (start, end, 
offsets, motif_name, strand, width)}
}
\description{
Given a set of motifs that form a window, it returns a dataframe with 
the data for the window (i.e, pastes motif names, offsets, strands, 
starts, ends in a string and calculates start and end) of the window.
}
\keyword{motifs}
\keyword{windows,}
