% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBedFile.R
\name{writeBedFile}
\alias{writeBedFile}
\title{writeBedFile}
\usage{
writeBedFile(df, IDs, fname)
}
\arguments{
\item{df}{input dataframe with "chrom", "chrom_start", "chrom_end"
and "strand" variables}

\item{IDs}{character vector with the same length as the number of rows
of the input dataframe, containing the interval names}

\item{fname}{output file name (.bed will be pasted to it)}
}
\description{
Creates a dataframe with bed-like file structure and writes it into 
a .bed file without the row names and the column names.
}
\keyword{bed}
